Generate types for every package
