+++ packages/embeds/embed-core/env.d.ts
@@ -1,7 +1,9 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
-  readonly NEXT_PUBLIC_WEBSITE_URL: string;
+  readonly EMBED_PUBLIC_WEBAPP_URL: string;
+  readonly EMBED_PUBLIC_VERCEL_URL: string;
+  readonly EMBED_PUBLIC_EMBED_LIB_URL: string;
 }
 
 interface ImportMeta {
