+++ packages/embeds/embed-core/package.json
@@ -1,10 +1,11 @@
 {
   "name": "@calcom/embed-core",
-  "version": "1.0.0",
+  "version": "1.1.0-beta.10",
   "description": "This is the vanilla JS core script that embeds Cal Link",
-  "main": "./index.ts",
+  "main": "./dist/embed/embed.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
-    "__build": "yarn tailwind && vite build",
+    "__build": "yarn tailwind && vite build && yarn tsc  --emitDeclarationOnly --declarationDir dist && cp -r ../../../apps/web/public/embed ./dist/",
     "__dev": "yarn __build --mode development",
     "build": "NEXT_PUBLIC_EMBED_FINGER_PRINT=$(git rev-parse --short HEAD) yarn __build",
     "build-preview": "PREVIEW_BUILD=1 yarn __build ",
@@ -16,8 +17,12 @@
     "type-check": "tsc --pretty --noEmit",
     "lint": "eslint --ext .ts,.js src",
     "embed-tests": "yarn playwright test --config=playwright/config/playwright.config.ts",
-    "embed-tests-quick": "QUICK=true yarn embed-tests"
+    "embed-tests-quick": "QUICK=true yarn embed-tests",
+    "prepare": "NEXT_PUBLIC_WEBAPP_URL='https://app.cal.com' yarn build"
   },
+  "files": [
+    "dist"
+  ],
   "postcss": {
     "map": false,
     "plugins": {
@@ -29,11 +34,10 @@
     "autoprefixer": "^10.4.7",
     "eslint": "^8.15.0",
     "postcss": "^8.4.13",
-    "vite": "^2.9.9"
-  },
-  "dependencies": {
+    "vite": "^2.9.9",
     "tailwindcss": "^3.0.24",
     "tsc": "^2.0.4",
     "typescript": "^4.6.4"
-  }
+  },
+  "dependencies": {}
 }
