+++ packages/embeds/embed-react/src/index.ts
@@ -1,4 +1,4 @@
-import { CalWindow } from "@calcom/embed-snippet";
+import { CalWindow } from "@calcom/embed-core";
 
 import Cal from "./Cal";
