+++ packages/embeds/embed-react/vite.config.js
@@ -2,10 +2,12 @@ import react from "@vitejs/plugin-react";
 import path from "path";
 import { defineConfig } from "vite";
 
+import viteBaseConfig from "../vite.config";
+
 // https://vitejs.dev/config/
 export default defineConfig({
+  ...viteBaseConfig,
   plugins: [react()],
-  envPrefix: "NEXT_PUBLIC_",
   build: {
     lib: {
       entry: path.resolve(__dirname, "src/index.ts"),
