+++ packages/embeds/embed-snippet/package.json
@@ -1,14 +1,23 @@
 {
   "name": "@calcom/embed-snippet",
-  "version": "1.0.0",
-  "main": "src/index.ts",
+  "version": "1.0.1-beta.10",
+  "main": "./dist/snippet.umd.js",
+  "module": "./dist/snippet.es.js",
   "description": "Vanilla JS embed snippet that is responsible to fetch @calcom/embed-core and thus show Cal Link as an embed on a page.",
   "scripts": {
-    "build": "vite build",
+    "build": "vite build && yarn tsc  --emitDeclarationOnly --declarationDir dist",
     "type-check": "tsc --pretty --noEmit",
-    "lint": "eslint --ext .ts,.js src"
+    "lint": "eslint --ext .ts,.js src",
+    "prepare":"NEXT_PUBLIC_EMBED_LIB_URL='https://app.cal.com/embed/embed.js' NEXT_PUBLIC_WEBAPP_URL='https://app.cal.com' yarn build"
   },
+  "files": [
+    "dist"
+  ],
+  "types": "./dist/index.d.ts",
   "devDependencies": {
     "eslint": "^8.15.0"
+  },
+  "dependencies": {
+    "@calcom/embed-core": "^1.1.0-beta.10"
   }
 }
