+++ packages/embeds/embed-snippet/vite.config.js
@@ -1,13 +1,11 @@
-require("dotenv").config({ path: "../../../.env" });
+import viteBaseConfig from "../vite.config";
 
 const path = require("path");
 const { defineConfig } = require("vite");
 
-process.env.NEXT_PUBLIC_VERCEL_URL = process.env.VERCEL_URL;
-
 module.exports = defineConfig({
+  ...viteBaseConfig,
   build: {
-    envPrefix: "NEXT_PUBLIC_",
     lib: {
       entry: path.resolve(__dirname, "src", "index.ts"),
       name: "snippet",
